Bucket Sort
Sorting Algorithms to sort n elements in a array using Bucket Sort.
Time Complexity :  If we assume that insertion in a bucket takes O(1) time then steps 1 and 2 of the above algorithm clearly take O(n) time. The O(1) is easily possible if we use a linked list to represent a bucket. Step 4 also takes O(n) time as there will be n items in all buckets.