Customised movie recommendation system using python
This is a single user system through which the user can maintain and filter movies and obtain Customized recommendations by sorting them according to film name, Director ,Genre ,Year ,Main actor ,IMDB rating and Duration. Also it allows the user to watch the movie trailer and the full movie straight through the software and also the watched status can be marked so that the user may keep note of the previously watched films. In