Beach House Problem
The main idea is to buy a big area of land in front of the beach. The plot can only be rectangular in shape. The beach is n units long and the length of the plot is i units. Given that the owner each has 1 unit. We need to find the maximum plot area in a rectangular shape which can be owned. The plot can also be made by combining two plots and by combining any fractional part of any land by any owner.
Input : 
           2
           5
           3 2 1 4 5
Output:
           8