Create JARVIS.py
This is a piece of Ai software that understands verbal
or written commands and completes task assigned by the
client. It is an example of weak Al that is it can only execute
and perform quest designed by the user. With the python
programming language, a script most commonly used by
the developers can be used to build your personal Al
assistant to perform task designed by the users.