Update leap year.py
Divisibility by 4: The primary rule is that a year is a leap year if it is divisible by 4. This is a simple rule that adds an extra day (February 29th) to the calendar every four years, helping to account for the fact that the Earth's orbit around the Sun takes approximately 365.2422 days.

Divisibility by 100: To further refine the system, years divisible by 100 are not leap years unless they are also divisible by 400. This rule removes some leap years, correcting for the fact that the basic rule of divisibility by 4 would make too many leap years. For example, without this rule, the year 1900 would be considered a leap year (divisible by 4), but it's not because it's divisible by 100 and not by 400.

Divisibility by 400: The exception to the above rule is that years divisible by 400 are still considered leap years. This rule ensures that certain years, like the year 2000, are leap years, even though they are divisible by 100.