Update binary-search.py
The earlier formula for finding mid includes high + low which has a problem. The problem is, in case if high + low has very large value that it will exceed the integer range for python while making the calculations, it will give error. So, applying the same formula but in a better way 
low + (high-low)/2 which is ultimately (high + low)/2