Greedy Algorithm to find Minimum number of Coins
assists in finding the minimum number of denominations needed to form a particular amount using a given set of coin denominations. The program should prompt the user to input the number of denominations, followed by the denominations in ascending order. After this, the program should ask the user for the target amount. Subsequently, it should compute and display the minimal number of coins needed to achieve the given amount, considering the available denominations.