Update main.js
In this the username from the href attribute of the anchor <a> element, but some contributor URLs may not follow the exact format you expect. Specifically, the URLs used in the href attribute might not always match "https://github.com/username", which can lead to undefined or incorrect URLs for the avatars.